[feat] category테이블의 region 크롤링
